+++ check_vps_btc.py
@@ -1,0 +1,11 @@
+import paramiko
+
+ssh = paramiko.SSHClient()
+ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+ssh.connect('147.93.156.165', port=22, username='root', password='rMM2m63P', timeout=10)
+
+# Search for BTCUSDT logs in the last 2000 lines
+cmd = 'journalctl -u cryptomentor.service -n 2000 --no-pager | grep BTCUSDT'
+stdin, stdout, stderr = ssh.exec_command(cmd)
+print(stdout.read().decode('utf-8'))
+ssh.close()